+++ retz-client/src/main/java/io/github/retz/web/MySocket.java
@@ -55,7 +55,7 @@ import java.util.concurrent.TimeoutException;
  *       the pair of CountDownLatch and bare String.
  */
 // NOTE: TODO: this message size limits the size of job, number of jobs listable via 'list' command
-@WebSocket(maxTextMessageSize = 64 * 1024)
+@WebSocket(maxTextMessageSize = 64 * 1024 * 1024)
 public class MySocket {
     private Session session;
     private CountDownLatch requestLatch;