+++ retz-common/src/main/java/io/github/retz/protocol/GetJobRequest.java
@@ -20,33 +20,16 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.util.List;
-
-public class ListJobResponse extends Response {
-    private List<Job> queue;
-    private List<Job> running;
-    private List<Job> finished;
+public class GetJobRequest extends Request {
+    private int id;
 
     @JsonCreator
-    public ListJobResponse(@JsonProperty("queue") List<Job> queue,
-                           @JsonProperty("running") List<Job> running,
-                           @JsonProperty("finished") List<Job> finished) {
-        this.queue = queue;
-        this.running = running;
-        this.finished = finished;
-    }
-    @JsonGetter("queue")
-    public List<Job> queue() {
-        return queue;
+    public GetJobRequest(@JsonProperty("id") int id) {
+        this.id = id;
     }
 
-    @JsonGetter("running")
-    public List<Job> running() {
-        return running;
-    }
-
-    @JsonGetter("finished")
-    public List<Job> finished() {
-        return finished;
+    @JsonGetter("id")
+    public int id() {
+        return id;
     }
 }