+++ retz-server/src/main/java/io/github/retz/scheduler/JobQueue.java
@@ -17,20 +17,31 @@
 package io.github.retz.scheduler;
 
 import io.github.retz.protocol.Job;
+import io.github.retz.protocol.StatusResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.*;
 import java.util.concurrent.BlockingQueue;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * On memory job queue to mediate job execution requests and resources
+ * TODO: make all these data tolerant against node or network failure
  */
 public class JobQueue {
+    private static final Logger LOG = LoggerFactory.getLogger(JobQueue.class);
+
     // CMT: I thought capital case was for static const, like in C++
     // REVIEW: static final fields should have CAPITAL_CASE name
     private static final BlockingQueue<Job> JOB_QUEUE = new LinkedBlockingDeque<>();
     private static final AtomicInteger COUNTER = new AtomicInteger(1);
+
+    private static final ConcurrentHashMap<String, Job> running = new ConcurrentHashMap<>(); // TaskID#getValue() -> Job
+    private static final ConcurrentLinkedDeque<Job> finished = new ConcurrentLinkedDeque<>();
 
     private JobQueue() {
     }
@@ -81,6 +92,25 @@ public class JobQueue {
         return ret;
     }
 
+    public synchronized static Optional<Job> getJob(int id) {
+        for(Job job : get()) {
+            if (id == job.id()) {
+                return Optional.of(job);
+            }
+        }
+        for(Map.Entry<String, Job> entry : running.entrySet()) {
+            if (id == entry.getValue().id()) {
+                return Optional.of(entry.getValue());
+            }
+        }
+        for(Job job : finished) {
+            if (id == job.id()) {
+                return Optional.of(job);
+            }
+        }
+        return Optional.empty();
+    }
+
     public static String now() {
         return Calendar.getInstance().getTime().toString();
     }
@@ -91,5 +121,41 @@ public class JobQueue {
 
     public static int size() {
         return get().size();
+    }
+
+    public static void start(String taskId, Job job) {
+        running.put(taskId, job);
+    }
+
+    public static void recoverRunning() {
+        for(Iterator<Map.Entry<String, Job>> it = running.entrySet().iterator(); it.hasNext();) {
+            Map.Entry<String, Job> entry = it.next();
+            try {
+                push(entry.getValue());
+            } catch (InterruptedException e) {
+                LOG.warn("Interruption: Job(id={}) must be rerun ({})", entry.getValue().id(), e.getMessage());
+                continue; // Avoid remove in case we could retry removal again
+            }
+            it.remove();
+        }
+    }
+
+    public static Job finish(String taskId) {
+        Job job = running.remove(taskId);
+        finished.add(job);
+        return job;
+    }
+
+    public static Map<String, Job> getRunning() {
+        return running;
+    }
+
+    public static void getAllFinished(List<Job> list, int limit) {
+        list.addAll(finished);
+    }
+
+    // Methods for test
+    public static void setStatus(StatusResponse response) {
+        response.setStatus(JobQueue.size(), running.size());
     }
 }