+++ retz-server/src/test/java/io/github/retz/web/WebConsoleTest.java
@@ -1,32 +1,33 @@
 /**
- *    Retz
- *    Copyright (C) 2016 Nautilus Technologies, KK.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
+ * Retz
+ * Copyright (C) 2016 Nautilus Technologies, KK.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package io.github.retz.web;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import io.github.retz.cli.FileConfiguration;
+import io.github.retz.protocol.*;
 import io.github.retz.scheduler.Applications;
 import io.github.retz.scheduler.JobQueue;
 import io.github.retz.scheduler.MesosFrameworkLauncher;
 import io.github.retz.scheduler.RetzScheduler;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import io.github.retz.protocol.*;
 import org.apache.commons.io.IOUtils;
 import org.apache.mesos.Protos;
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -96,8 +97,10 @@ public class WebConsoleTest {
     @Test
     public void list() throws Exception {
         assert webClient.connect();
-        ListJobResponse res = (ListJobResponse) webClient.list();
+        ListJobResponse res = (ListJobResponse) webClient.list(64);
         assert res.queue().isEmpty();
+        assert res.running().isEmpty();
+        assert res.finished().isEmpty();
         webClient.disconnect();
     }
 
@@ -123,7 +126,7 @@ public class WebConsoleTest {
             Application app = res.applicationList().get(0);
             assert app.getAppid().equals("foobar");
             System.out.println("================================" + res.applicationList());
-            for(Application a: res.applicationList()){
+            for (Application a : res.applicationList()) {
                 System.out.println("=========== app:" + a);
                 System.out.println("=========== app:" + a.getAppid());
                 System.out.println("=========== app:" + a.getFiles());
@@ -150,6 +153,9 @@ public class WebConsoleTest {
 
             maybeJob = JobQueue.pop();
             assert !maybeJob.isPresent();
+
+            GetJobResponse getJobResponse = (GetJobResponse) webClient.getJob(235561234);
+            Assert.assertTrue(!getJobResponse.job().isPresent());
         }
 
         {
@@ -175,8 +181,11 @@ public class WebConsoleTest {
             assert sres.job.id() >= 0;
             System.err.println(sres.job.scheduled());
 
-            ListJobResponse listJobResponse = (ListJobResponse) webClient.list();
+            ListJobResponse listJobResponse = (ListJobResponse) webClient.list(64);
             assert listJobResponse.queue().size() == 1;
+
+            GetJobResponse getJobResponse = (GetJobResponse) webClient.getJob(sres.job.id());
+            Assert.assertEquals(sres.job.cmd(), getJobResponse.job().get().cmd());
 
             maybeJob = JobQueue.pop();
             assert maybeJob.isPresent();